Conflicts:
	pom.xml
	presto-kafka/src/main/java/com/facebook/presto/kafka/KafkaInternalFieldDescription.java
	presto-kafka/src/main/java/com/facebook/presto/kafka/KafkaRecordSet.java
	presto-kafka/src/main/java/com/facebook/presto/kafka/KafkaRecordSetProvider.java
	presto-kafka/src/main/java/com/facebook/presto/kafka/KafkaSimpleConsumerManager.java
	presto-kafka/src/main/java/com/facebook/presto/kafka/KafkaSplitManager.java

+++ presto-spi/src/main/java/com/facebook/presto/spi/block/BlockUtil.java
@@ -59,7 +59,7 @@ final class BlockUtil
     static int calculateNewArraySize(int currentSize)
     {
         // grow array by 50%
-        long newSize = currentSize + (currentSize >> 1);
+        long newSize = (long) currentSize + (currentSize >> 1);
 
         // verify new size is within reasonable bounds
         if (newSize < DEFAULT_CAPACITY) {
